+++ pre-processing/primod/ribametamod/ribametamod.py
@@ -238,7 +238,7 @@ class RibaMetaMod:
             column="node_id",
         )
 
-        basin_ids = np.unique(self.ribasim_model.basin.profile["node_id"])
+        basin_ids = np.unique(self.ribasim_model.basin.profile.df["node_id"])
         missing = ~np.isin(self.basin_definition["node_id"], basin_ids)
         if missing.any():
             missing_basins = self.basin_definition["node_id"].to_numpy()[missing]
