+++ tests/test_ribamod_cases.py
@@ -7,8 +7,8 @@ from primod.ribamod import DriverCoupling, RibaMod
 def create_basin_definition(
     ribasim_model: ribasim.Model, buffersize: float
 ) -> gpd.GeoDataFrame:
-    node = ribasim_model.node.static
-    basin_nodes = ribasim_model.basin.static["node_id"].unique()
+    node = ribasim_model.database.node.df
+    basin_nodes = ribasim_model.basin.static.df["node_id"].unique()
     basin_geometry = node.loc[basin_nodes].geometry
     # Call to_numpy() to get rid of the index
     basin_definition = gpd.GeoDataFrame(
